[Interop] WPT and fix for click dispatch with chorded buttons.
This CL adds a WPT for click-like events from chorded buttons, which
is currently being discussed in PEWG as an Interop problem:
  https://github.com/w3c/pointerevents/issues/530

The CL also fixes a behind-a-flag crack in Chromium.  The crack
prevented firing of click-like events with the chorded buttons when
ClickToCapturedPointer is enabled.  This is because the call to
`DispatchMouseClickIfNeeded` is wrongly skipped in this case since
the `pointer_event->type()` is `kPointermove` not `kPointerup`.

Bug: 40851596
Change-Id: I11606ed8f709f08aaa2d8aba3ebaeb07f26bbc1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6059955
Reviewed-by: Robert Flack <flackr@chromium.org>
Auto-Submit: Mustaq Ahmed <mustaq@chromium.org>
Commit-Queue: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395404}