Use MockPermissionPromptFactory for WebXr tests
In preparation for implementing android_browsertests and as a benefit/
cleanup to the existing windows versions of the tests, this adds
infrastructure to appropriately create and use a
MockPermissionPromptFactory during WebXr tests.

Note that in order to simplify the lifecycle requirements for the
test creators, the MockPermissionPromptFactory now extends the
PermissionRequestManager::Observer interface so that it can properly
delete it's manager when it is deleted.

Bug: 381000093
Change-Id: I4c31fe39e5d29dd085a61c58c67452f832faada7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6063412
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Florian Jacky <fjacky@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395406}