[Blink] Unify CanvasResource accelerated TransferableResource prep paths
ExternalCanvasResource's method for preparing a TransferableResource is
now identical to that used for other accelerated CanvasResource
subclasses (as can be seen in this CL by looking at
the PrepareAcceleratedTransferableResourceFromClientSI() implementation
and comparing it to the
PrepareAcceleratedTransferableResourceWithoutClientSI() code that is
being removed). This CL removes ExternalCanvasResource's custom code and
has it go through the common codepath.

A couple notes wrt the two minor differences that exist between those
two codepaths:

* The DCHECK(SharedGpuContext::IsGpuCompositingEnabled()); line
  presumably holds for ExternalCanvasResource just as it presumably
  holds for the other type of accelerated CanvasResources that go
  through this codepath (as the software compositor would in any case
  just drop a GPU TransferableResource on the floor).
* ExternalCanvasResource always uses accelerated raster [1].

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/graphics/canvas_resource.h;l=484?q=ExternalCanvasResource::UsesA&sq=&ss=chromium

Bug: 352263194
Change-Id: I1646c9a8cb7296565793cdb8cd8f311813871363
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6088115
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395422}