[passage-embeddings] Service + TabHelper to compute passages embeddings
EmbedderTabHelper is a stripped down HistoryEmbeddingsTabHelper with
no dependency on HistoryService or HistoryEmbeddingsService. It also
waits for a configurable amount of time after all tabs finish loading.

EmbedderService has a similar interface to history_embeddings::Embedder
and is made a profile-keyed service to encapsulate the queuing
and scheduling logic prior to invoking the Embedder model.

EmbedderService is gated by passage_embeddings::kPassageEmbedder and
the the EmbedderTabHelper is gated by history embeddings opt-out so
the opted in population does not end up embedding additional passages.

Bug: 383380610
Low-Coverage-Reason: EXPERIMENTAL_CODE
Change-Id: I26332ffd77a32b924572c85b572fa608b24e6520
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6084617
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395455}