Fix calling trim() on undefined in PrintPreviewMarginControlElement
In the Print Preview code, margin_control.html has an on-change handler
that calls PrintPreviewMarginControlElement's onInputChange_() method.
When this gets called, there is no CustomEvent detail, so the call fails
because onInputChange_() is expecting a string in the detail field.
Avoid this problem by removing the on-change handler.

Change-Id: I746e5a159b9aa4beb786cd971cc21723557d81b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6081402
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395462}