[BabelOrca] Re-factor caption translator for ease of use
Re-factors the BabelOrca caption translator to set the languages
of the translation per call to translate, rather than using an
init translation and set callback method.  This simplifies the
API for the translator  and keeps the location of where the
source and target language of the translation are set in the
same place that the call to translate happens.

TESTED: on brya, passes current test coverage.
Bug: 377696975,377544063
Change-Id: Id5e775226c9cd39a9a6ef00858b67bd06c0e29af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6051572
Reviewed-by: Ahmed Nasr <anasr@google.com>
Commit-Queue: Avynn Donaghe <avynn@google.com>
Cr-Commit-Position: refs/heads/main@{#1395463}