a11y: Set ToggleButton initial checked state
The a11y information for the checked state of ToggleButton is updated
when the state of the button changes. However, since no initial state is
set, if e.g. Quick Settings is opened and we attempt to read the toggle
state it will be kNone even though the toggle exists, is visible, and is
off. This change sets the initial state.

Bug: 366781447
Test: Added unit test.
Change-Id: Iafebf8d464540d2017c078538485806d12033720
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6088006
Reviewed-by: Dirk Pranke <dpranke@google.com>
Commit-Queue: Chad Duffin <chadduffin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395532}