[PIX] Remove FacilitatedPaymentsTypes variant from histograms
Most histograms have FacilitatedPaymentsTypes as the placeholder in the
name. This was done in anticipation that eWallets would reuse these
metrics as they are. But that's not the case, eWallets have additional
params in histograms.

Hardcode "Pix" in Pix histograms. If eWallets also log the same
histogram, then the FacilitatedPaymentsTypes variant can be used. It's a
no-op.

Bug: 377126728, 367751320
Change-Id: Idfea8db6ac3988d227283a5eae4a5f0ea633d050
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6086567
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Vishwas Uppoor <vishwasuppoor@google.com>
Cr-Commit-Position: refs/heads/main@{#1395563}