Add ability for App-Bound to signal need for reencryption
App-Bound decrypt now returns an additional success code if it
determines that re-encryption might be needed for decrypted
data e.g. the key has been rotated.

If this success code is seen by the client code then an
EncryptData call is made on the same COM thread as the Decrypt
with the Microsoft::WRL::ComPtr held active to ensure the
service stays alive for both calls.

The newly encrypted data is returned to the caller and the
caller should then persist the newly encrypted data.

The App-Bound key provider is updated to persist this data to
the pref in this case.

Tests are added to verify all this behavior including adding
a 'fake reencrypt' switch to the elevated service that can be
used during testing. New tests are created for the App-Bound
key provider as these did not exist previously.

Metrics are also added to measure the success and frequency of
the re-encrypt operations.

Finally, the whole re-encrypt is placed behind a disabled by
default feature flag to ensure this can be rolled out safely.

BUG=383157187,382234536

Change-Id: Ia599fde9e83cc66e4daaff637ba16b60580a67cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6084659
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Xiaoling Bao <xiaolingbao@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395573}