Add histogram to track time for first accessibility focus in Android
This CL tracks the amount of time it takes for a single instance of
WebContentsAccessibilityImpl to start its native initialization, until
the same instance first gives a node accessibility focus. We only
track from native initialization onward because that is the first
signal that an AT is running. We restrict the recording to TalkBack,
since we know that it always tries to focus the root node, but other
ATs may not ever send an accessibility focus action, which would
skew the metrics. We only record this once per instance, since there
can only be one first focus, and we set the scale to 5 seconds max
with the expectation of changing this if needed after initial data
gives a sense of an accurate scale for the value.

AX-Relnotes: N/A
Bug: 383387881
Change-Id: I5d3a1b1ba834975b5f5c96d612cde873248c14f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6090216
Reviewed-by: Lucas Radaelli <lucasradaelli@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1395580}