[Cast Streaming] Remove deprecated EmitFrames method
This patch removes the EmitFrames method from the VideoEncoder API,
since it is no longer used.

This method was added in https://codereview.chromium.org/863083002 for
use with the now-deleted H264VideoToolboxEncoder. This was the only
encoder that implemented this method, and since it has been deleted this
method can also be deleted.

Bug: 282984511
Change-Id: I4dc6718f20fc1956b412f1e45af1d4205d770c36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6089521
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395608}