Fix crash in MagicBoostCardController::ShowOptInUi
MagicBoostCardController::ShowOptInUi() can be called when right-click
on pdf and on the browser, and this can happen simultaneously since the
pdf one happens via an async callback. Thus, we need to remove the
CHECK, and close the previous opt in widget if needed.

Bug: 383628875
Change-Id: I88abd2c3c26235abd9f234c9c769d209021bb5cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6088248
Commit-Queue: Andre Le <leandre@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Auto-Submit: Andre Le <leandre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395628}