[FedCM] Do not show continue button if account is filtered
Currently we show the continue button when there is a single newly
logged in filtered account to be displayed in the UI. This CL fixes this
by ensuring that a filtered account cannot be marked as selected. It
also adds checks so that we only consider the single account chooser to
be on when the single account being shown is not filtered out.

Fixed: 383804523
Change-Id: Ic734a06a8865085991bb83751a3ab99575aaf311
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6091939
Reviewed-by: Zachary Tan <tanzachary@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Nicolás Peña <npm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395671}