Add support for performance hints when loading on-device model
This allows passing a performance hint to the backend when loading the
on-device model. This can be used to tune various params in the backend
or use a smaller model if requested.

This is controlled from the optimization guide side by a new feature
param "compatible_low_tier_on_device_performance_classes", which
allows us to choose a set of performance classes that will use the
kFastestInference performance hint.

As part of this change, all the feature params related to model
performance were moved under a new base::Feature to allow them to be
adjusted independent of the other on-device related params.

Bug: 379723772
Change-Id: I28db8de699a8af026ee2ee1cfa08136342860e13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6089104
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395676}