HttpStreamPool: Ensure Group is destroyed after SSLConfig change
If a Group and a corresponding AttemptManager only have in-flight
attempts and there are no requests (jobs) nor idle/handed-out streams,
the group and the attempt manager should be destroyed after cancelling
in-flight attempts. Before this CL there was a case where we didn't
destroy them when an SSLConfig change happened and there were only
in-flight attempts:
-> (SSLConfig change event)
-> HttpStreamPool::OnSSL{,ForServers}Changed()
-> HttpStreamPool::Group::Refresh()
-> HttpStreamPool::AttemptManager::CancelInFlightAttempts()

Ensure that these are destroyed if these can be completed after
cancelling in-flight attempts.

Bug: 346835898
Change-Id: I106ce206af1d4f08bb3350776826b0c734fffb34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6090630
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395689}