[MediaRouter] Update OpenScreenListener class
This CL updates OpenScreenListener to enable it to discovery Open
Screen protocol based services. This is the first step to integrate
Open Screen protocol into MediaRouter and the discovered services are
not exposed to MediaRouter yet.

The next step is to provide an class named OpenScreenMediaRouterProvide
as a bridge between MediaRouter and OpenScreenListener.

Bug: none
Change-Id: I06e08737b4a887b49e615e497d1082b5e6f6c732
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6069867
Commit-Queue: Wei4 Wang <wei4.wang@intel.com>
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395700}