[Task manager] Show the tab name on renderer tasks
The task manager gets updated when navigation finishes and changes the
task name to the name of the tab.

We enable the WebContentsTaskProvider, modulo the code related to the
following features that are unavailable on android chrome.

* extensions
* background contents
* guest view

The change in tab_helpers.cc is necessary because the tab model android
chrome uses is different from the one other platforms use, which calls
CreateForTabContents in tab_features.cc.

This change increases the number of mutable constants usage in chrome on
android, because of the kTagKey in web_contents_tag. This will be fixed
separately.

Fixed: 379192565
Bug: 383245212
Binary-Size: Will be fixed separately (see above).
Change-Id: Ibe8689288334be65fc8e6ce35e81ff2ac7804efa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6087495
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395709}