coral: fix the crash on removing the chip by user
After making the BirchBarController as a BirchCoralProvider::Observer,
removing a group will notify the controller to remove corresponding
item. However, when a chip is removed by user, the observer will also
be called which causes duplicated removing.

The cl fixed the issue by first moving the removing item out of the
item list of controller such that the observer knows the the item
is being removed to avoid duplicated removing.

Test: Manual, CoralControllerTest.NoCrashOnRemovingChipByUser
Bug: b:383781526
Change-Id: I675a98f4835fed764b362643ef29407a515eebdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6090264
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Xiaodan Zhu <zxdan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395721}