sunfish: Disable launcher nudge by default in tests
This fixes many launcher pixel tests and interactive UI tests after
adding a field trial config for Scanner, which will cause the launcher
nudge to appear in tests.

Manually enable the launcher nudge in the Sunfish tests which need it.

This CL is based on the similar disable-for-test methods in
`AppListNudgeController`.

Bug: b:381965299
Change-Id: I6a6a2100f3a84fe74c1592a61e7aac0b2fa3b7fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6085937
Reviewed-by: Elijah Hewer <hewer@chromium.org>
Auto-Submit: Michael Cui <mlcui@google.com>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Michael Cui <mlcui@google.com>
Cr-Commit-Position: refs/heads/main@{#1395749}