Update the default value of has_web_request_api_proxy
In crbug.com/362539771, we added the fix not to dispatch SWAutoPreload
feature to the navigation which involves WebRequest API, and that is
currently implemented as a feature param.

However, since the partner wants to check this behavior on their side,
we'd like to change the default value of this feature param. The feature
param is under the Finch control, so we can safely change the default
value.

Bug: 383841345
Change-Id: I8310ed720f91986f6d1c3c220874dd7fe86e68b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6088252
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Keita Suzuki <suzukikeita@chromium.org>
Commit-Queue: Keita Suzuki <suzukikeita@chromium.org>
Auto-Submit: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395788}