url_matcher: Rename AddFilters to AddFiltersWithLimit
Rename AddFilters to AddFiltersWithLimit and add a max_filters
parameter to explicitly control the number of added filters.
This is a preparatory refactoring that does not introduce any
logic changes; the existing filter limit remains in effect.

This change sets the stage for a follow-up update where the
filter limit will be increased for the URLBlocklist and
URLAllowlist policies.

Bug: 380396936
Change-Id: I042098c46bc13b336000a8cfa00b67e16c8307bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6088218
Commit-Queue: Emmanuel Arias Soto <eariassoto@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395848}