Allow persisting extension logs in Kiosk
Done via the chrome.systemLog API. Extensions can already use the API,
but the logs are going to the Device Event Log by default (not
persisted in between sessions).

Currently, if the IsDeviceExtensionsSystemLogEnabled policy is enabled
and you are in one of the allowed session types (login screen and MGS),
the logs are persisted. Add Kiosk to the allowed session types.

Bug: b:370612770
Change-Id: I548e52550f6aacc5aabf96485abe50cdb3be3839
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6075072
Commit-Queue: Maria Petrisor <mpetrisor@chromium.org>
Reviewed-by: Polina Bondarenko <pbond@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395858}