Respect KeyboardDefaultToFunctionKeys policy during CRD session
This was broken in M113 in https://crrev.com/c/4288536.

Before that change the KeyboardDefaultToFunctionKeys policy was
applied globally for all input devices, but after that change the
policy is applied to each input device separately, and the code
instead checks the `top_row_are_fkeys` field inside the
`KeyboardSettings`.

Which shouldn't have caused any issues, except for the fact that CRD
uses a fake device id (ED_REMOTE_INPUT_DEVICE) for which there are no
`KeyboardSettings`, causing input through CRD to fall back to the
default `return false` statement at the end of the function,
irrespective of the policy's value.

Fixed: 381153238
Test: New unittest + manually deployed
Change-Id: I375e1e55daa748d9318af6ed436753a2a4be3b35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6088918
Commit-Queue: Jeroen Dhollander <jeroendh@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395882}