[ios] Fix AppMetricsAppStateAgent for multi-profile
First split the code responsible for tracking the session duration
to a ProfileAgent. There will be one instance per ProfileState and
each of them will be responsible for tracking the session for that
profile.

This requires moving the storage for the start of the session from
the AppState to the agent itself (there is no need to store it in
the ProfileState since it is only used by the agent).

Also fixes bug where the AppMetricsAppStateAgent assumed that the
SceneState events and AppState/ProfileState initialisation stage
had an expected ordering (this is not true, they can mostly happen
in any relative order).

Fixed: 381849395
Change-Id: I06974c26500b1921f619a7bce3cb702926e479c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6088912
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395900}